+++ tests/rwlock6_t2.c
@@ -86,9 +86,9 @@ main()
   pthread_t wrt1;
   pthread_t wrt2;
   pthread_t rdt;
-  int wr1Result = 0;
-  int wr2Result = 0;
-  int rdResult = 0;
+  intptr_t wr1Result = 0;
+  intptr_t wr2Result = 0;
+  intptr_t rdResult = 0;
   struct _timeb currSysTime;
   const DWORD NANOSEC_PER_MILLISEC = 1000000;
